requirements.txt
pygame>=2.5.0
pyyaml>=6.0
numpy>=1.24.0

pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
safety>=2.3.0
bandit>=1.7.0
psutil>=5.9.0